chore(with-storage-sync): move with-storage-sync files
