+++ libs/ngrx-toolkit/src/index.ts
@@ -23,7 +23,7 @@ export {
   withStorageSync,
   IndexedDBSyncConfig,
   SyncConfig,
-} from './lib/storage-sync/with-storage-sync';
+} from './lib/with-storage-sync';
 export { withImmutableState } from './lib/immutable-state/with-immutable-state';
 export { withFeatureFactory } from './lib/with-feature-factory';
 export { withConditional, emptyFeature } from './lib/with-conditional';
